Fix login URL for teams with SSO
Slack teams have the option to turn on login via Google auth, which results in the inital page not including a login form.  Adding the query-string parameter `no_sso` gets you to the actual login page.  I'm making the assumption that including that parameter does not break the loing for teams that _don't_ have Google SSO turned on.